requirements.txt
Pillow
pytesseract
opencv-python
opencv-python-headless
argparse